+++ tests/_all/line_continuation.golden
