--- tests/_sh/line_continuation.c
@@ -1,19 +1,0 @@
-#include <stdio.h>
-
-void main() {
-
-/**/
-int foo = 0;
-
-/\
-*
-*/ fo\
-o +\
-= 0\
-x\
-10\
-200;
-
-    printf("%d", foo);
-    return 0;
-}
