--- tests/_sh/line_continuation.golden
@@ -1,1 +1,0 @@
-66048
